Fix checks for static busybox
Bug fix: In order to avoid file not found errors in the chroot, busybox should be statically linked, so an additional check is carried out. This requires the 'file' command to be present, though 'ldd' could also be used. Ideally, the package (ebuild) dependencies would specify the 'static' USE flag for busybox, but this should not be assumed.